requirements.txt
torch>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
mediapipe>=0.10.0
opencv-python>=4.8.0
gtts>=2.4.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
numpy>=1.24.0

